+++ package/Editor/XRBuildProcessorHelper.cs
@@ -31,24 +31,7 @@ namespace UnityEditor.XR.Management
 
         void CleanOldSettings()
         {
-            UnityEngine.Object[] preloadedAssets = PlayerSettings.GetPreloadedAssets();
-            if (preloadedAssets == null)
-                return;
-
-            var oldSettings = from s in preloadedAssets
-                where s.GetType() == typeof(T)
-                select s;
-
-            if (oldSettings.Any())
-            {
-                var assets = preloadedAssets.ToList();
-                foreach (var s in oldSettings)
-                {
-                    assets.Remove(s);
-                }
-
-                PlayerSettings.SetPreloadedAssets(assets.ToArray());
-            }
+            BuildHelpers.CleanOldSettings<T>();
         }
 
         void SetSettingsForRuntime(UnityEngine.Object settingsObj)